// src/generated/endpoints.ts
export const endpoints = {
// ...
	"DELETE /nodes/{node}/qemu/{vmid}": {
		method: "DELETE",
		path: "/nodes/{node}/qemu/{vmid}",
		name: "destroy_vm",
		description: "Destroy the VM and all used/owned volumes. Removes any VM specific permissions and firewall rules",
		allowToken: true,
		rootOnly: false,
		protectedCall: true,
		proxyTo: "node",
		returnType: "string",
		params: {
			"destroy-unreferenced-disks": { type: "boolean", optional: true, default: 0 },
			node: { type: "string", optional: false, format: "pve-node" },
			purge: { type: "boolean", optional: true },
			skiplock: { type: "boolean", optional: true },
			vmid: { type: "integer", optional: false, format: "pve-vmid" },
		},
	},
// ...
	"POST /nodes/{node}/lxc": {
		method: "POST",
		path: "/nodes/{node}/lxc",
		name: "create_vm",
		description: "Create or restore a container.",
		allowToken: true,
		rootOnly: false,
		protectedCall: true,
		proxyTo: "node",
		returnType: "string",
		params: {
// ...
			features: { type: "string", optional: true, propertyString: true },
			force: { type: "boolean", optional: true },
			"ha-managed": { type: "boolean", optional: true, default: 0 },
			hookscript: { type: "string", optional: true, format: "pve-volume-id" },
			hostname: { type: "string", optional: true, format: "dns-name" },
			"ignore-unpack-errors": { type: "boolean", optional: true },
			lock: { type: "string", optional: true, enum: ["backup", "create", "destroyed", "disk", "fstrim", "migrate", "mounted", "rollback", "snapshot", "snapshot-delete"] },
// ...
			rootfs: { type: "string", optional: true, propertyString: true },
// ...
		},
	},
// ...
	"POST /nodes/{node}/lxc/{vmid}/move_volume": {
		method: "POST",
		path: "/nodes/{node}/lxc/{vmid}/move_volume",
		name: "move_volume",
		description: "Move a rootfs-/mp-volume to a different storage or to a different container.",
		allowToken: true,
		rootOnly: false,
		protectedCall: true,
		proxyTo: "node",
		returnType: "string",
		params: {
			bwlimit: { type: "number", optional: true, default: "clone limit from datacenter or storage config" },
			delete: { type: "boolean", optional: true, default: 0 },
// ...
		},
	},
// ...
	"POST /nodes/{node}/spiceshell": {
		method: "POST",
		path: "/nodes/{node}/spiceshell",
		name: "spiceshell",
		description: "Creates a SPICE shell.",
		allowToken: true,
		rootOnly: false,
		protectedCall: true,
		proxyTo: "node",
		returnType: "null",
		params: {
			cmd: { type: "string", optional: true, enum: ["ceph_install", "login", "upgrade"], default: "login" },
			"cmd-opts": { type: "string", optional: true, default: "" },
			node: { type: "string", optional: false, format: "pve-node" },
			proxy: { type: "string", optional: true, format: "address" },
		},
	},
// ...
} as const satisfies Readonly<Record<string, EndpointInfo>>